# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "EventEase API Documentation",
    "description": "Multi-Vendor Event Management System API Documentation (Auto-Generated)",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "description": "Register a new user to the national portal",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "email", "password"],
              "properties": {
                "name": { "type": "string", "example": "Ayesha Bibi" },
                "email": { "type": "string", "example": "ayesha@example.com" },
                "password": { "type": "string", "example": "password123" },
                "role": { "type": "string", "example": "client" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created successfully" },
          "400": { "description": "User already exists or bad validation" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "description": "Authenticate user and return access token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": { "type": "string", "example": "ayesha@example.com" },
                "password": { "type": "string", "example": "password123" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Invalid login credentials" }
        }
      }
    },
...
  }
}